data-sync/consumer/requirements.txt
qdrant-client>=1.9.1
kafka-python>=2.0.2
loguru>=0.7.2
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.3.0
huggingface-hub>=0.23.0
numpy>=1.26.4
pymongo>=4.7.2
